+++ ArkavQuarium/src/com/ArkavQuarium/Guppy.java
@@ -9,10 +9,10 @@ public class Guppy extends Fish {
     private int foodCount;
     private int price;
     private int foodThreshold;
-    private long lastCoinTime;
-    private long coinInterval;
+    private double lastCoinTime;
+    private double coinInterval;
 
-    public Guppy(long time, double x, double y) {
+    public Guppy(double time, double x, double y) {
         super(time, x, y);
 
         growthStep = 1;
@@ -39,16 +39,16 @@ public class Guppy extends Fish {
         this.foodCount = foodCount;
     }
 
-    public long getLastCoinTime() {
+    public double getLastCoinTime() {
         return lastCoinTime;
     }
 
-    public void setLastCoinTime(long lastCoinTime) {
+    public void setLastCoinTime(double lastCoinTime) {
         this.lastCoinTime = lastCoinTime;
     }
 
     @Override
-    public void moveRandom(long time, Point max) {
+    public void moveRandom(double time, Point max) {
         double moveX, moveY, newRad;
 
         if (isTimeToChangeDirection(time)) {
@@ -80,7 +80,7 @@ public class Guppy extends Fish {
         setY(getY() + moveY);
     }
 
-    public void moveToEat(LinkedList<Food> listfood, Food food, long time) {
+    public void moveToEat(LinkedList<Food> listfood, Food food, double time) {
         if (food.findDistance(new Point(getX(), getY())) <= speed) {
             setX(food.getX());
             setY(food.getY());
@@ -104,7 +104,7 @@ public class Guppy extends Fish {
         }
     }
 
-    public boolean isTimeToGiveCoin(long time) {
+    public boolean isTimeToGiveCoin(double time) {
         boolean waktunya = (time - lastCoinTime) >= coinInterval;
         if (waktunya) { lastCoinTime = time; }
         return waktunya;
